Make make_small_change reliable for small files
For very small files, where only 1 or 2 bytes is changed, there is a
realistic chance that the new random couple bytes would happen to be
equal to the original bytes, and the file wouldn't actually change.

This is why the test_update_many_files test was flaky. It was making 64
one-byte changes. Each had a 1 in 256 chance of ending up with the same
file. Multiply that by 64 files to change, and there is a 1 in 4 chance
that the test will have one identical file and fail. That's what I was
seeing. The test would fail about 1 in 4 times.

This change enhances the make_small_change function to double-check that
it is actually writing different data for those small chunk sizes. This
makes the test reliable again.