Verify repo on commit fail and then verify
In the situation where a commit fails, but is intact, we also want to
verify the repo. This change tweaks the control flow so that that
happens. Before, the commit verifier was suppressing the exception and
then the repo verifier wasn't running.

This also adds unit tests to the RepoVerifier and fixes a bug in the
CommitVerifier.