Fix compile on Rust 1.17 and 1.18
The updated macro systems really do not like my reference to `$crate` in
the error definition macros. This code compiles if we remove it though
and access the `ObjectKey` type by absolute path, `::dag::ObjectKey`.

This will only be valid inside the crate, but I think that's ok because
we are calling a macro here, not exporting one.